move requirement check into a class
